requirements.txt
openai
llama_index
pypdf
gradio
langchain
panel
python-dotenv
chromadb
loguru
